# swagger.yaml
openapi: 3.0.3
info:
  title: API de Promessas de Campanha
  version: 1.0.0
  description: |
    API para cadastro e consulta de candidatos e suas promessas de campanha.
servers:
  - url: http://localhost:3000
paths:
...
  /api/v1/promises/{promiseId}:
    patch:
      summary: Atualiza status/andamento de uma promessa
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: promiseId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePromiseInput'
      responses:
        '200':
          description: Promessa atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Promise'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Promessa não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token ausente ou inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/promises/{promiseId}/comments:
    post:
      summary: Adiciona um comentário ao andamento
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: promiseId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePromiseCommentInput'
      responses:
        '201':
          description: Comentário criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromiseComment'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Promessa não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token ausente ou inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/offices:
    get:
      summary: Lista cargos disponíveis
      parameters:
        - in: query
          name: type
          required: false
          schema:
            $ref: '#/components/schemas/ElectionType'
          description: Filtra cargos pelo tipo de eleição (FEDERAL_ESTADUAL ou MUNICIPAL)
      responses:
        '200':
          description: Lista de cargos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Office'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Cria um novo cargo
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOfficeInput'
      responses:
        '201':
          description: Cargo criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Office'
        '400':
          description: Dados inválidos ou nome duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token ausente ou inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/offices/{officeId}:
    patch:
      summary: Atualiza um cargo
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: officeId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOfficeInput'
      responses:
        '200':
          description: Cargo atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Office'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Cargo não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token ausente ou inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Promise:
      type: object
      properties:
        id:
          type: integer
        candidate_id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/PromiseStatus'
        progress:
          type: integer
          minimum: 0
          maximum: 100
        comments_count:
          type: integer
          example: 3
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        comments:
          type: array
          items:
            $ref: '#/components/schemas/PromiseComment'
      required:
        - id
        - candidate_id
        - title
        - status
        - progress
        - created_at
        - updated_at
        - comments
    PromiseComment:
      type: object
      properties:
        id:
          type: integer
        promise_id:
          type: integer
        content:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - promise_id
        - content
        - created_at
...
    UpdatePromiseInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/PromiseStatus'
        progress:
          type: integer
          minimum: 0
          maximum: 100
    CreatePromiseCommentInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...
    Office:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          $ref: '#/components/schemas/ElectionType'
      required:
        - id
        - name
        - type
    CreateOfficeInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
      required:
        - name
    UpdateOfficeInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
...
    PromiseStatus:
      type: string
      enum:
        - NOT_STARTED
        - IN_PROGRESS
        - COMPLETED
    ElectionType:
      type: string
      description: Tipo de eleição/cargo
      enum:
        - FEDERAL_ESTADUAL
        - MUNICIPAL
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Erro interno do servidor
      required:
        - error
  responses:
    InternalError:
      description: Erro interno inesperado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'